install and configure karma, sauce labs and travis-ci
All of the work on integrating rollup into the build process for
distribution files, created some friction when trying to integrate
Karma. Due to better support, I've added Browserify to the test
preprocessors for Karma, and anticipate refactoring the gulp build
tasks with it as well.

In fact, the "mocha in browser" approach seemed to fall flat once
I looked at introducing Sauce Labs for cross browser testing...
and now I understand where Karma fits into the ecosystem!

Unfortunately, it means some pretty dramatic refactoring had to
happen, undoing some of the past month's work... and there's still
more refactoring to be done.

That said, everything appears to be working locally with Karma and
Sauce Labs, and the next step is confirming that Travis CI is
setup correctly alongside the two.