add constructor unit test
Dang, code coverage is awesome! I noticed that the instantiated
object was not being tested when ScrollRreveal isn't supported.

I made a few changes to be able to stub the method checking for
support, and also to the noop instance so that it still has noop
public methods.

The noop instance should prevent runtime errors on browsers that
don't support CSS Transition and Transform. One could check that
the `sr` class has been added to `<html>`, but this approach
allows the developer to be optimistic.